+++ openrouteservice/src/main/java/heigit/ors/services/routing/requestprocessors/gpx/GpxRoutingResponseWriter.java
@@ -14,18 +14,18 @@ import heigit.ors.routing.WeightingMethod;
 import heigit.ors.services.routing.RoutingServiceSettings;
 import heigit.ors.util.AppInfo;
 import heigit.ors.util.GeomUtility;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.BoundsType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.CopyrightType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.EmailType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.Gpx;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.GpxExtensions;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.LinkType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.MetadataType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.PersonType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.RteType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.RteTypeExtensions;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.WptType;
-import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.WptTypeExtensions;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.BoundsType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.CopyrightType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.EmailType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.Gpx;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.GpxExtensions;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.LinkType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.MetadataType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.PersonType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.RteType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.RteTypeExtensions;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.WptType;
+import heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans.WptTypeExtensions;
 
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
