--- openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/gpxUtil/beans/EmailType.java
@@ -1,91 +1,0 @@
-
-package heigit.ors.util.GlobalExportWriter.gpxUtil.beans;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * 
- * 	 An email address.  Broken into two parts (id and domain) to help prevent email harvesting.
- *
- *
- * <p>Java class for emailType complex type.
- *
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * {@code
- * <complexType name="emailType">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <attribute name="id" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
- *       <attribute name="domain" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "emailType")
-public class EmailType {
-
-    @XmlAttribute(name = "id", required = true)
-    protected String id;
-    @XmlAttribute(name = "domain", required = true)
-    protected String domain;
-
-    /**
-     * Gets the value of the id property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getId() {
-        return id;
-    }
-
-    /**
-     * Sets the value of the id property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setId(String value) {
-        this.id = value;
-    }
-
-    /**
-     * Gets the value of the domain property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getDomain() {
-        return domain;
-    }
-
-    /**
-     * Sets the value of the domain property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setDomain(String value) {
-        this.domain = value;
-    }
-
-}
