--- openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/gpxUtil/beans/package-info.java
@@ -1,2 +1,0 @@
-@javax.xml.bind.annotation.XmlSchema(namespace = "https://raw.githubusercontent.com/GIScience/openrouteservice-schema/master/gpx/v1/ors-gpx.xsd", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package heigit.ors.util.GlobalExportWriter.gpxUtil.beans;
