--- openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/gpxUtil/exportWriter/GpxExportWriter.java
@@ -1,4 +1,0 @@
-package heigit.ors.util.GlobalExportWriter.gpxUtil.exportWriter;
-
-public class GpxExportWriter {
-}
