+++ openrouteservice/src/main/java/heigit/ors/util/GlobalResponseProcessor/gpxUtil/beans/package-info.java
@@ -1,0 +1,2 @@
+@javax.xml.bind.annotation.XmlSchema(namespace = "https://raw.githubusercontent.com/GIScience/openrouteservice-schema/master/gpx/v1/ors-gpx.xsd", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+package heigit.ors.util.GlobalResponseProcessor.gpxUtil.beans;
