+++ openrouteservice/src/main/java/heigit/ors/util/GlobalResponseProcessor/gpxUtil/exportWriter/GpxResponseWriter.java
@@ -1,0 +1,4 @@
+package heigit.ors.util.GlobalResponseProcessor.gpxUtil.exportWriter;
+
+public class GpxResponseWriter {
+}
